+++ finagle-zipkin-core/src/main/scala/com/twitter/finagle/zipkin/Flags.scala
@@ -1,0 +1,6 @@
+package com.twitter.finagle.zipkin
+
+import com.twitter.app.GlobalFlag
+import com.twitter.finagle.zipkin.core.Sampler
+
+object initialSampleRate extends GlobalFlag[Float](Sampler.DefaultSampleRate, "Initial sample rate")